refactor(table-data-grid): split fetch composables
