+++ packages/core/table-data-grid/src/composables/index.ts
@@ -1,6 +1,5 @@
 import useI18n from './useI18n'
 import { useTableDataGridConfig } from './useTableDataGridConfig'
-import { useTableDataGridFetch } from './useTableDataGridFetch'
 import { useDatatableColumnDefs } from './useDatatableColumnDefs'
 import { useDatatableColumnSizing } from './useDatatableColumnSizing'
 import { useDatatablePagination } from './useDatatablePagination'
@@ -8,14 +7,15 @@ import { useDatatableSelection } from './useDatatableSelection'
 import { useTableDataGridColumnLayoutSync } from './useTableDataGridColumnLayoutSync'
 import { useTableDataGridConfigSync } from './useTableDataGridConfigSync'
 import { useTableDataGridGridLifecycle } from './useTableDataGridGridLifecycle'
+import { useTableDataGridInfiniteFetch } from './useTableDataGridInfiniteFetch'
 import { useTableDataGridInteractions } from './useTableDataGridInteractions'
+import { useTableDataGridPaginationFetch } from './useTableDataGridPaginationFetch'
 import { useTableDataGridRefreshTriggers } from './useTableDataGridRefreshTriggers'
 import { useTableDataGridState } from './useTableDataGridState'
 
 // All composables must be exported as part of the default object for Cypress test stubs
 export default {
   useTableDataGridConfig,
-  useTableDataGridFetch,
   useDatatableColumnDefs,
   useDatatableColumnSizing,
   useDatatablePagination,
@@ -23,7 +23,9 @@ export default {
   useTableDataGridColumnLayoutSync,
   useTableDataGridConfigSync,
   useTableDataGridGridLifecycle,
+  useTableDataGridInfiniteFetch,
   useTableDataGridInteractions,
+  useTableDataGridPaginationFetch,
   useTableDataGridRefreshTriggers,
   useTableDataGridState,
   useI18n,
