+++ packages/core/table-data-grid/tsconfig.build.json
@@ -6,6 +6,7 @@
   "exclude": [
     "src/**/*.cy.ts",
     "src/**/*.spec.ts",
+    "src/test-utils",
     "sandbox",
     "dist"
   ]
